reorganize msms scripts; add exclusion list template
