+++ msms_autox_generator/etc/1536spot.csv
