+++ msms_autox_generator/etc/24spot.csv
