+++ msms_autox_generator/etc/384spot.csv
