+++ msms_autox_generator/etc/48spot.csv
