+++ msms_autox_generator/etc/6144spot.csv
