+++ msms_autox_generator/etc/96spot.csv
