+++ msms_autox_generator/etc/exclusion_list_template.csv
@@ -1,0 +1,2 @@
+m/z
+379.15
