+++ msms_autox_generator/etc/geometry_list.txt
