+++ msms_autox_generator/etc/ms1_autox_generator.cfg
