+++ msms_autox_generator/gui.py
@@ -7,8 +7,8 @@ from lxml import etree as et
 from pymaldiproc.data_import import import_mzml, import_timstof_raw_data
 from pymaldiproc.preprocessing import align_spectra, get_feature_matrix
 from pymaldiviz.util import *
-from bin.layout import *
-from bin.util import *
+from msms_autox_generator.layout import *
+from msms_autox_generator.util import *
 from dash import State, callback_context, no_update, dash_table, MATCH, ALL
 from dash_extensions.enrich import Input, Output, DashProxy, MultiplexerTransform, Serverside, ServersideOutputTransform
 import dash_bootstrap_components as dbc
