+++ setup.py
@@ -12,16 +12,16 @@ if os.path.isfile('requirements.txt'):
 
 setup(
     name='flex_maldi_dda_automation',
-    version='0.4.0-alpha',
+    version='0.4.0b0',
     url='https://github.com/gtluu/flex_maldi_dda_automation',
     license='Apache License',
     author='Gordon T. Luu',
     author_email='gtluu912@gmail.com',
-    packages=['bin', 'etc'],
+    packages=['msms_autox_generator'],
     include_package_data=True,
     package_data={'': ['*.cfg']},
-    description='MALDI-DDA automation script for timsTOF fleX AutoXecute datasets',
-    entry_points={'console_scripts': ['process_maldi_dda=bin.run:main']},
+    description='timsTOF fleX MALDI AutoXecute Automation Scripts',
+    #entry_points={'console_scripts': ['process_maldi_dda=msms_autox_generator.run:main']},
     install_requires=install_requires,
     setup_requires=install_requires
 )
